refactor(update): clarify managed handoff ownership
