+++ src/infra/update-startup.test.ts
@@ -36,7 +36,6 @@ const {
     pid: 12345,
     command: "openclaw update --yes --channel beta --timeout 2700",
     logPath: "/tmp/openclaw-handoff.log",
-    ownsHandoff: true,
   })),
 }));
 
@@ -260,7 +259,6 @@ describe("update-startup", () => {
       pid: 12345,
       command: "openclaw update --yes --channel beta --timeout 2700",
       logPath: "/tmp/openclaw-handoff.log",
-      ownsHandoff: true,
     });
     resetUpdateAvailableStateForTest();
   });
@@ -1161,12 +1159,11 @@ describe("update-startup", () => {
     mockNpmChannelTag("beta", "2.0.0-beta.1");
     detectRespawnSupervisorMock.mockReturnValue("launchd");
     startManagedServiceUpdateHandoffMock.mockResolvedValueOnce({
-      status: "started",
+      status: "joined",
       pid: 12345,
       command: "openclaw update --yes --channel beta --timeout 2700",
       logPath: "/tmp/openclaw-handoff.log",
       handoffId: "handoff-existing",
-      ownsHandoff: false,
     });
 
     await runAutoUpdateCheckWithDefaults({
